Add basic support for templates.
Pybind11 does not yet support passing unique pointers into function arguments so the first test in templates_test.py always fails for now. (NEEDS_SMART_HOLDER)

As a team we decided to use a stop-gap method of checking the template name to bypass a blocking issue in the CLIF matcher (b/118736768).

PiperOrigin-RevId: 357008053