Add Bazel BUILD files for //clif/backend/... and //clif/protos/...
This CL adds a WORKSPACE file that loads and configures CLIF dependencies needed for backend and protos targets. clif_deps.bzl and clif_configure.bzl provide functions to properly configure the repository, which can be reused by 3rd-party consumers of CLIF to simplify WORKSPACE setup.

Currently, the following dependencies are loaded:
- abseil-cpp (https://github.com/abseil/abseil-cpp)
- llvm (https://github.com/google/llvm-bazel)
- googletest (https://github.com/google/googletest)
- protobuf (https://github.com/protocolbuffers/protobuf)

BUILD rules are also updated via buildozer to compile properly in the OSS version. bazel build ... and bazel test ... are passing.

Note that matcher_test needs to be updated because there isn't a trivial way to pass absolute paths to clif backend directory via Bazel, which some of the tests seem to expect.

PiperOrigin-RevId: 358483027