Make cindex extractor support checking overloaded functions.
Functions with parameters like reference to pointers, rvalue references, pointer to pointers, pointer to constants might not work properly.

This CL introduces very low-level manipulations of C++ type names (stripping spaces, stars, ampersands) that are difficult to generalize and maintain. We are exploring alternative approaches that may allow us to avoid such manipulations. In the meantime this CL unblocks the work on the pybind11 code generator.

PiperOrigin-RevId: 362354347