Fixing Python reference count leaks in `PyObjRef`.
Follow-on to cl/360248907 that added a test for the leaks.

Positive:

Note that PyCLIF currently has the best known solution for the combination of
  (virtual function overrides implemented in Python)
x (passing `shared_ptr`/`unique_ptr` back from Python to C++).
The leaks fixed by this CL appear to be unfortunate simple oversights.
The core related mechanisms are very well thought out.
In particular, for passing `shared_ptr` the PyCLIF approach
[here](https://github.com/google/clif/blob/13f60ba90ce113ab2c359b67734589860f25bfae/clif/python/runtime.h#L123)
is somewhat similar to the [open pybind11 PR #2839](https://github.com/pybind/pybind11/pull/2839).
For passing `unique_ptr`, the
[`HoldPyObj`](https://github.com/google/clif/blob/13f60ba90ce113ab2c359b67734589860f25bfae/clif/python/runtime.cc#L37)
logic used in conjunction with disowning (see generated code example below)
currently has no equivalent in pybind11, even in the pybind11
[`smart_holder`](https://github.com/pybind/pybind11/tree/smart_holder) branch.

Issues:

Global testing of earlier versions of this CL surfaced one test failure.

To root-cause this failure, virtual_py_cpp_mix_test.py was added. That in
turn led to the collateral discovery of a 3rd bug, see the `NOT GOOD` comment
in virtual_py_cpp_mix_test.py. This needs to be worked on later/separately.

The critical new test in virtual_py_cpp_mix_test.py is
`testGetFromCppUniquePtrPassingCppDerived`. It does not reproduce the
global test failure exactly, but initially it

* reproducibly ran to completion when NOT using ASAN.
* reproducibly generated a `heap-buffer-overflow` when running with ASAN.

Inspection of the PyCLIF-generated Python C API code revealed that this `reinterpret_cast`
```
  auto& shared = reinterpret_cast<clif_testing_python_virtual__py__cpp__mix_clifwrap::pyBase::wrapper*>(py)->cpp;
```
is improper (undefined behavior): it casts from a `pyCppDerived::wrapper`, which does not have an `Overrider` to a `pyBase::wrapper`, which does have an `Overrider`.
Resolving this properly is more work than we want to invest in the Python C API code generator at this stage of the pybind11 integration, but this CL implements a workaround that makes all tests run ASAN clean, including the one test that failed before in the global testing. The workaround has two parts:

* Using a `dynamic_cast` (see types.py) as a way to find out if `ThisPtr` is a `PyObjRef`, which means the wrapper has an `Overrider`. If the `dynamic_cast` succeeds, the resulting pointer can be used for the `HoldPyObj` logic. Otherwise it is certain that this logic is not needed because the type does not have an `Overrider`.
* Reordering the multiple-inheritance order for the `Overrider` (see gen.py), to move the wrapped C++ type first, `PyObjRef` second. This fixed the one test that failed before in the global testing, ASAN clean.

The result of the improper `reinterpret_cast` is still being used for the `Detach` step, which disowns the Python object to a `unique_ptr`. This would certainly need fixing if we had plans to use the Python C API code generator in the long run, or if there had been more than the one failing test, but in the current situation, given that all tests run ASAN clean, and that the improper cast has been in use for a very long time, it does not seem to be a wise investment of effort. The issue is now marked prominently in the generated code (and therefore also in the code generator), as a permanent reminder, until the code is removed after the pybind11 integration is completed.

For easy reference, this is the PyCLIF-generated Python C API code for passing `shared_ptr`/`unique_ptr` from Python to C++:
```
bool Clif_PyObjAs(PyObject* py, std::shared_ptr<::smart_ptrs::Operation>* c) {
  CHECK(c != nullptr);
  ::smart_ptrs::Operation* cpp = clif_testing_python_smart__ptrs_clifwrap::pyOperation::ThisPtr(py);
  if (cpp == nullptr) return false;
  auto& shared = reinterpret_cast<clif_testing_python_smart__ptrs_clifwrap::pyOperation::wrapper*>(py)->cpp;
  *c = ::clif::MakeSharedVirtual<::smart_ptrs::Operation>(shared, py);
  return true;
}

bool Clif_PyObjAs(PyObject* py, std::unique_ptr<::smart_ptrs::Operation>* c) {
  CHECK(c != nullptr);
  ::smart_ptrs::Operation* cpp = clif_testing_python_smart__ptrs_clifwrap::pyOperation::ThisPtr(py);
  if (cpp == nullptr) return false;
  auto& shared = reinterpret_cast<clif_testing_python_smart__ptrs_clifwrap::pyOperation::wrapper*>(py)->cpp;
  shared->HoldPyObj(py);
  if (!shared.Detach()) {
    shared->DropPyObj();
    PyErr_SetString(PyExc_ValueError, "Cannot convert Operation instance to std::unique_ptr.");
    return false;
  }
  c->reset(cpp);
  return true;
}
```

Collateral code health improvement: guaranteed initialization of `PyObjRef` member variables (pointers). In the google3 environment there is no actual problem, but externally a mix of `NDEBUG` defined/undefined is possible and could cause undefined behavior. The initializations also guard against surprising bugs while refactoring.

PiperOrigin-RevId: 363020182