Make pybind11 code generator support un-property and pointer attributes.
Previously we generate `def_readwrite` for raw_pointer and unique_ptr attributes. But this is unsupported. Now we generate `def_property` with customized getter and setter for these cases. When the attribute is unique_ptr, we will also take ownership of the Python object.

Also refactored classes.py so that we do not need to iterate through the ast multiple times.
Also deleted unused code in utils.py.
Also modified the order of function parameters to make them consistent with others.
Also split nested_fields_test to multiple test cases because the behavior on unique_ptr parameters is different between pybind11 and PyCLIF.

PiperOrigin-RevId: 372387891