Removing obsolete ` (maybe related to >>> go/clif_h_infix <<< change?)` hint.
It is now more likely to be confusing than helpful.

The modified line raising the `NameError` is not covered by tests, but I tested it manually by triggering the exception via a small temporary change elsewhere. I also used codesearch to search for "go/clif_h_infix": the only match is the code removed here.

PiperOrigin-RevId: 375147875