Enable use of C++ fully-qualified names for `capsule`.
Immediate motivation for this change: unblock cl/379850888.

This just requires a one-line change in pytd2proto.py, to turns off a simple check.

Unfortunately the unit test cannot easily be fit into t9_test, because the
* t9 test is using a `t9` namespace,
* therefore the `capsule` C++ name must start with `::`,
* but then the C API code generator produces invalid code.

To not disturb more than necessary for this very small (but critical) production code change, a new unit test is added, with one function in the global namespace, mimicking exactly what is needed to unblock cl/379850888.

This CL supports a Large-Scale Change (LSC): go/swig-python-exit-2020-lsc

PiperOrigin-RevId: 380363311