BEGIN_PUBLIC Adding `weakref` support for PyCLIF `class`es. END_PUBLIC
This change supports the ongoing work on migrating pywraprpc from SWIG to PyCLIF (b/185378877).

An attempt was made to also extend the `weakref` support to PyCLIF iterator types, but there were issues. Currently this feature is not needed, therefore it was systematically omitted (see `if iterator` logic in gen.py).

Note that the pybind11 tests in iterator_test.py do not actually run. This is an existing bug uncovered while working on this CL; it is being investigated separately. The expectation was that the pybind11 version will support weakrefs for iterators, which is the primary reason for not spending time on this feature for the C API version.

An earlier version of this CL changed the generated `wrapper` ABI, which led to two TGP test failures. These are tracked under b/196881887. The `wrapper` members are now arranged for backward ABI compatibility.

PiperOrigin-RevId: 391337906