Removing most Python 2 related code from PyCLIF.
This is a 1st-stage cleanup with a firehose, based on systematically searching for `PY_.*VER`, `py2`, `py3`, and replacing/removing all obviously obsolete code. The details are messy, but the unit tests not directly related to the C API code generator are unchanged (with one small, obviously completely safe exception: third_party/clif/testing/t10.h), and a Global TAP presubmit was run.

b/198606402 was created to track follow-on cleanup of Python 2 compatibility code in third_party/clif/python/types.h.

Other Python 2 releated code missed at this 1st stage will be cleaned up as it is discovered; this is expected to be of much smaller scale.

PiperOrigin-RevId: 395999002