Make pybind11 code generator partially support capsules.
Capsules defined in other .clif files are not supported.

This is achieved by generating a lambda expression when a function takes in or returns a capsule. When a function returns a capsule, we return a `CapsuleWrapper<Payload>` object. When a function takes a capsule, we make the function takes a `CapsuleWrapper<Payload>` object. The type caster of CapsuleWrapper object is defined in runtime.h.

This change is to support //thread/python:thread, which is a dependency for lots of Python libraries.

PiperOrigin-RevId: 404000612