Generate appropriate bindings code for const decls.
Previously we generate `def_readonly_static` for const decls in classes. This is incorrect when exposing enums to their parent scope. See the modified t3.clif in the cl.

Also do not generate `.export_values()` for enums. Based on https://pybind11.readthedocs.io/en/stable/classes.html?highlight=export_values#enumerations-and-internal-types, this is unnecessary for C++11 and newer versions.

PiperOrigin-RevId: 404390256