`@extend` support for nested classes.
Caveat: type names in extended functions or properties of nested classes must be fully qualified. For example:

```
class Outer:
  class Inner:
    def foo(other: Inner)  # Not using @extend: OK with unqualified type name.

    @extend
    def bar(other: Outer.Inner)  # Needs fully-qualified type name.
```

Unfortunately this is not easy to fix (cost/benefit is high).

This CL is in support of pending cl/403208037 (pywrapgslbchannel SWIG-to-CLIF migration).

This CL was fully TGP-tested twice:

* Initially using `_CLIF_TO_PYI_FROM_SOURCE` here: https://source.corp.google.com/piper///depot/google3/devtools/python/clif_to_pyi/clif_to_pyi.bzl;l=65;bpv=1;bpt=0;rcl=396915646 (http://tap/OCL:406236281:BASE:406711074:1635745498642:898e4421).

* Then again after updating clif_to_pyi.par using update_par.py (TAP URL below).

The most direct and cleanest way to ensure full consistency is to update clif_to_pyi.par together with the PyCLIF changes in this CL.

PiperOrigin-RevId: 406898595