Fixing `@enable_instance_dict` base-derived inconsistencies.
This is the simplest approach and ensures full consistency: the `instance_dict` pointer always exists, even though it is used only if `@enable_instance_dict` is specified.

Significant effort would be required to avoid the extra pointer: the inheritance graph needs to be searched, which is especially difficult if base and derived classes are in different .clif files.

The switch to pybind11 will make the effort obsolete.

It is not even certain that the extra pointer leads to a detectable increase in memory consumption, in the context of a typical application.

We know for sure though that the attempt to optimize away the pointer has caused substantial confusion and loss of time (b/203381676).

PiperOrigin-RevId: 407115019