Remove unnecessary parameter preprocessing when generating lambda expressions.
Previously the code generator emulates the logic of legacy CLIF. But most of the logic is unnecessary when generating pybind11 code. Legacy CLIF creates new instances of parameters. For example, the following exists in the generated code of legacy CLIF:

```
Type arg1;
if (!Clif_PyObjAs(a[0], &arg1)) return ArgError("Func", names[0], "Type", a[0]);
```

Therefore legacy CLIF needs to generate different code based on whether the parameter type  has default constructor, has public destructor, etc.. But for pybind11, because it does not create new instances of parameter types, most of the logic is unnecessary. But we still need to so preprocessing for abstract types, because abstract type as function parameter is not allowed.

PiperOrigin-RevId: 414580473