Generate operator bindings code instead of lambda expressions for operators.
Sometimes operators are also qualified for lambda expression generation. But the generated code is invalid because operators are special. It implicitly use `self` as the first parameter. But when generating lambda expressions, we will add another parameter if the function is a member function.

The newly added test case fails without this change.

PiperOrigin-RevId: 429612742