Break out stltypes_fwd.h to fix newly added statusor_variant_test
* https://yaqs.corp.google.com/eng/q/4541161143598579712

The newly added statusor_variant_test intentionally exercises `absl::variant` in addition to `std::variant`, although the former is now an alias for the latter in google3. This is to catch any issues in other environments (e.g. OSS or b/215649709). It is safest to keep the `absl::variant` tests around as long as possible, until the alias is purged.

Notes regarding TGP testing:

* The TGP for this CL (TESTED URL below) has 148 "possibly failing or broken" targets.
* The TGP for a completely unrelated CL has the exact same 148 "possibly failing or broken" targets: http://tap/OCL:446456389:BASE:446720765:1651763788191:19200d5a
* Therefore it is certain that all 148 broken/failed targets are unrelated to this CL.
* See also: Description of cl/446492469, which was similarly plagued by a large number of unrelated failures.

PiperOrigin-RevId: 446801735