Support passing absl::Status as function argument.
This is for `set_util_status` defined at https://source.corp.google.com/piper///depot/google3/net/rpc/python/rpc.clif;l=135. The reason I do not directly implement this in `third_party/pybind11_abseil` is to avoid dependency with `util/task/python`. Whenever there is a function argument with type `absl::Status`，I generate a lambda expression and convert the function argument to `PyCLIFStatus<absl::Status>`.

`requires_status` is to make sure we are not wrapping types called `Status` in CLIF but is not `absl::Status` in C++.

`absl::Status*` as function argument is not supported.

PiperOrigin-RevId: 454239621