Add `HasTemplateParameter` to `// CLIF use` comments.
It is hard for the code generator (implemented in Python) to tell whether a C++ type requires template parameters, so we modify the `// CLIF use` comments to include this information.

Note that I did not remove `NumTemplateParameter` because it already exists in google3: https://source.corp.google.com/search?q=%22NumTemplateParameter:%22%20lang:cc.

PiperOrigin-RevId: 491727471