Remove test for very low-value feature blocking PyCLIF-pybind11 work.
Using `unique_ptr` to pass an enum type is a very artificial construct.

Supporting this feature in PyCLIF-pybind11 is possible, but not worth the effort, including the effort for maintaining the added complexity.

It seems very unlikely that this feature is used in client code, and if it is, it will be best to change the client code to pass the enum directly (less generated machine code, more performant).

Tested:

```
tap_presubmit -p clif.pybind11 --detach --email -c 495090196
```

* http://tap/OCL:495090196:BASE:495090752:1670962252251:52d68751

PiperOrigin-RevId: 495125268