Do not disable SWIG/Clif interoperability for `@final` classes.
I'm not 100% sure what the intent was here, but it looks like `Clif` uses `as_MyType` to convert a SWIG object to a Clif object.
When a class is marked `final`, this disables support for the conversion, which I think should still be supported as is does not actually involve inheritance as far as I can tell.

Example cl that breaks without this change: cl/500681149

PiperOrigin-RevId: 501297368