Fix long-standing ref-count stealing bug in `clif::callback`.
See b/282776731 for full background, especially b/282776731#comment3 & b/282776731#comment4.

The old behavior is bug-prone and definitely not something we want to introduce into the pybind11 ecosystem, in connection with go/pyclif_pybind11_fusion.

The final TGP has only "1 possibly already failing or broken" target, which is definitely unrelated to this CL.

PiperOrigin-RevId: 532595246