`LOG(ERROR)` Python exception raised in C++ callback.
Adding the `LOG(ERROR)` is only a small improvement over the status quo, but makes it easier to experiment with `LOG(FATAL)`. There is also a slight chance that the `LOG(ERROR)` will bring previously overlooked issues to the attention of teams running production systems.

OSS version tested interactively in a docker container.

PiperOrigin-RevId: 534992892