Revamp `clif::callback` implementation (minor behavior changes only).
Unfortunately this PyCLIF-C-API cleanup is still necessary, to facilitate converging the behavior with PyCLIF-pybind11. The existing PyCLIF-C-API behavior is very unsafe, therefore not suitable for the pybind11 ecosystem. This CL sets the stage for follow-on convergence work: cl/534979529 changes 3 `LOG(ERROR)` introduced here to `LOG(FATAL)`; see b/284164589 for the motivation.

The main objective of this CL is to clearly and comprehensively report the 4 distinct situations that can lead to Python exceptions when `clif::callback::Func::operator()` is called:

* Low-level Python error (allocating tuple for callback arguments failed).

* C++ argument to-python conversion failure.

* Python exception raised in callback.

* Return value from-python conversion failure.

This CL also fixes undesirable behavior of the `ArgIn` implementation that went unnoticed before:

* google3/third_party/clif/python/stltypes.h;l=342;rcl=534992892

There are two flaws:

* The loop for processing of arguments continues while the Python error indicator is set (i.e. `PyErr_Occurred()` is true).

* Minor: The `Py_CLEAR(*a);` there is a no-op because `*a` is known (from the context) to be the `nullptr` anyway.

This CL fixes both problems.

For completeness, the trigger for this work was:

```
blaze test //consensus/config/python:basic_config_clif_test
```

The utilstatus_test addition in this CL was reduced from that test.

Also TGP-tested in combination with cl/534979529 (http://OCL:536217365:BASE:536289904:1685423096058:1ec70728). Note that cl/534979529 has multiple blockers that need to be submitted first.

There is no solution yet to enable changing the only remaining `LOG(ERROR)` to `LOG(FATAL)`. This is still blocked by `//consensus/config/python:basic_config_clif_test`.

PiperOrigin-RevId: 536385762