Fix oversights: Python error not set in `Clif_PyObjAs()` functions.
Oversights introduced with cl/538823846, cl/539993368.

The oversights are compensated for in the generated (C-API) code, for example:

Before this CL:
```
TypeError: PassMapInt() argument v is not valid for ::std::map<int, int> (list instance given)
```

With this CL:
```
TypeError: PassMapInt() argument v is not valid for ::std::map<int, int> (list instance given): list instance is not convertible to a std::map-like object
```

This CL is to make the conversion functions more consistently reusable (especially in connection the PyCLIF-pybind11 development work).

It was systematically verified that all inserted `SetIsNotConvertibleError()` calls are covered by unit tests. Tests were added as needed. However, there are no assertions on the exact error messages, for compatibility with PyCLIF-pybind11 testing.

PiperOrigin-RevId: 564462441