Document nullptr and None
I'm actually not sure if this is correct, so please double check this. Based on the test in cl/572986800 it seems correct to me.

I couldn't find any CLIF documentation for how nullptr and None work. I was confused on cl/572986800 how a function saying
-> LoginTicket
could return None. I would have thought that would require
-> LoginTicket | None
or
-> Optional[LoginTicket]

In fact, I don't see a single .clif file using | None or Optional[] :
https://source.corp.google.com/search?q=lang:clif%20%5C%7C.*None
https://source.corp.google.com/search?q=lang:clif%20Optional%5C%5B

PiperOrigin-RevId: 575853650