Fix the reference count stealing problem in `clif::callback`.
This is a follow up of cl/532595246. When the first parameter of the callback function is not `PyObject*`, and the second parameter of the callback function is `PyObject*`, the newly added function template specialization for `PyObject*` is not executed. See b/282776731#comment6 for details.

TGP only includes unrelated failures: https://fusion2.corp.google.com/presubmit/568959836/OCL:568959836:BASE:569532419:1696009662955:2e10d708/targets.

PiperOrigin-RevId: 576333623