Import google/pywrapcc from GitHub (#30034, #30087, #30088).
For easy direct access:

* https://github.com/google/pywrapcc/pull/30034

* https://github.com/google/pywrapcc/pull/30087

* https://github.com/google/pywrapcc/pull/30088

Note regarding the change in std_containers_copy_move_test.py:

When the test was added (cl/570839777), the undocumented expectation was that it is hyper-sensitive, but that it will be adjusted or made more permissive (similar to e.g. google3/third_party/clif/testing/python/return_value_policy_test.py;l=25;rcl=534200687) as needed.

#MIGRATION_3P_PYBIND11__DEFAULT

```
  - 24a3b1c3729401ca661efacfbbd13a83117e1bae Add `case return_value_policy::_clif_automatic` in type_c... by Ralf W. Grosse-Kunstleve <rwgk@google.com>
  - 015834b13a99d233c15406c36a3a44d27ebfc846 Change `array_caster` in pybind11/stl.h to support value ... by Ralf W. Grosse-Kunstleve <rwgk@google.com>
  - ce00785ef877f0bb1dba10115f00366111583b3c Add `release_gil_before_calling_cpp_dtor` annotation for ... by Ralf W. Grosse-Kunstleve <rwgk@google.com>
```

PiperOrigin-RevId: 599883612