Use newly added pybind11 `__setstate__[non-constructor]` feature to replace ad hoc workaround.
See the description of https://github.com/google/pywrapcc/pull/30094 for background.

Note for completeness: This CL is an improvement in its own right, but the original trigger was to remove a blocker for child cl/600951417 (Use the default Python `type` as `pybind11::metaclass` for PyCLIF-wrapped classes). The ad hoc workaround replaced in this CL does not work with the default Python `type` as metaclass.

PyCLIF-pybind11 TGPs using this change: b/287289622#comment54

TGP testing of this CL in isolation is not required: The only TAP project exercising this change is clif.pybind11.

PiperOrigin-RevId: 602505662