Remove the `output_to_genfiles = True` parameter from Starlark rules (go/lsc-output-to-genfiles).
This will cause the rule to create its outputs in `blaze-bin` (the default) instead of `blaze-genfiles`.

Outputting to `blaze-genfiles` is only required for backwards compatibility with other rules or tools that make assumptions about the location of output files. In the long term, all uses of `blaze-genfiles` will be replaced with blaze-bin. See go/output_to_genfiles for more information.

Cases where `output_to_genfiles = True` is accompanied by a comment justifying its presence have been excluded from this LSC. If this CL breaks you, we will approve a rollback, but expect you to also send a CL documenting the parameter, to facilitate future cleanup efforts. Please indicate the specific rule or tool you need to be compatible with, so that we know what additional work needs to be done.

Tested:
    TAP --sample ran all affected tests and none failed
    http://test/OCL:616233458:BASE:616161007:1710535883404:a8cc69cd
PiperOrigin-RevId: 616621130