Fixed a greedy regex that was matching spaces inside of the macro's data.
This change modifies the regex so that the name matches only up to the first space, so that spaces within the macro's definition are kept as part of the definition. The old broken behavior caused a bug where the name would match up to the last space within the definition, making the name include part of the actual definition and the matched definition only contain the tail of the actual definition.

This error affects the use of `interface` declarations in `.clif` files, when those declarations have functions with comments, and when those `.clif` files are referenced by other `.clif` files, even when the consuming library does not attempt to reuse an `interface` declaration. It would have instead crashed while parsing the generated C++ header.

cl/663079816 provides an example usage of `interface` that triggers this bug and is fixed by this CL.

PiperOrigin-RevId: 666477200